+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/termbg/compare/v0.4.0...Unreleased) - ReleaseDate
 
+* [Added] emacs detection
+
 ## [v0.4.0](https://github.com/dalance/termbg/compare/v0.3.0...v0.4.0) - 2022-01-25
 
 * [Breaking Changed] drop Terminal::RxvtCompatible